Version 1.1
This version is more stable. Let me explain, Firefox ESR is now going to be the new browser for the Raspberry Pi on Raspbian. It's going to be slow at the start, but it's one of the applications allowing autostart so it can run headless. You may want to allow access for Firefox to do this on the Raspbian. Next, the sleep timer is going to 60 this time so you have enough time to move around outside of your pi. Other than that, this is a more stable version than V1.0.

Next update: Creating a system where the "no motion" adds up to a certain number to detect when someone really isn't at their desk. This is to stop entirely the moving around all the time system when you're listening to music.