Version 1.3
New updates make this version more stable. A new redesign of the code made the program a little bit cleaner. The program now relies on a 50 increment counter and is reset once the 50 increment counter is reached. The program also will reset the counter if motion is detected at any time. This way, it avoids the reliance of motion detection once.

This update focused more on creating an increment counter to make the system more reliable.